Move disaggregated mode into disagg context
