+++ dbms/src/Flash/Mpp/MPPHandler.cpp
@@ -17,6 +17,7 @@
 #include <Flash/Mpp/MPPHandler.h>
 #include <Flash/Mpp/Utils.h>
 #include <Interpreters/Context.h>
+#include <Interpreters/SharedContexts/Disagg.h>
 
 #include <ext/scope_guard.h>
 
@@ -33,7 +34,7 @@ namespace
 void addRetryRegion(const ContextPtr & context, mpp::DispatchTaskResponse * response)
 {
     // For tiflash_compute mode, all regions are fetched from remote, so no need to refresh TiDB's region cache.
-    if (!context->isDisaggregatedComputeMode())
+    if (!context->getSharedContextDisagg()->isDisaggregatedComputeMode())
     {
         for (const auto & table_region_info : context->getDAGContext()->tables_regions_info.getTableRegionsInfoMap())
         {
