+++ dbms/src/Storages/Transaction/LearnerRead.cpp
@@ -18,6 +18,7 @@
 #include <Common/TiFlashMetrics.h>
 #include <Flash/Coprocessor/DAGContext.h>
 #include <Interpreters/Context.h>
+#include <Interpreters/SharedContexts/Disagg.h>
 #include <Poco/Message.h>
 #include <Storages/DeltaMerge/ScanContext.h>
 #include <Storages/Transaction/KVStore.h>
@@ -148,7 +149,7 @@ LearnerReadSnapshot doLearnerRead(
     const LoggerPtr & log)
 {
     assert(log != nullptr);
-    RUNTIME_ASSERT(!(context.isDisaggregatedComputeMode() && context.useAutoScaler()));
+    RUNTIME_ASSERT(!(context.getSharedContextDisagg()->isDisaggregatedComputeMode() && context.getSharedContextDisagg()->use_autoscaler));
 
     auto & tmt = context.getTMTContext();
